# PageStorage: Log down the gc time for external pages