--- a/dbms/src/Storages/DeltaMerge/StoragePool.cpp
+++ b/dbms/src/Storages/DeltaMerge/StoragePool.cpp
@@ -544,10 +544,6 @@ void StoragePool::dataRegisterExternalPagesCallbacks(const ExternalPageCallbacks
         break;
     }
     case PageStorageRunMode::ONLY_V3:
-    {
-        data_storage_v3->registerExternalPagesCallbacks(callbacks);
-        break;
-    }
     case PageStorageRunMode::MIX_MODE:
     {
         // We have transformed all pages from V2 to V3 in `restore`, so
@@ -570,13 +566,10 @@ void StoragePool::dataUnregisterExternalPagesCallbacks(NamespaceId ns_id)
         break;
     }
     case PageStorageRunMode::ONLY_V3:
-    {
-        data_storage_v3->unregisterExternalPagesCallbacks(ns_id);
-        break;
-    }
     case PageStorageRunMode::MIX_MODE:
     {
-        // no need unregister callback in V2.
+        // We have transformed all pages from V2 to V3 in `restore`, so
+        // only need to unregister callbacks for V3.
         data_storage_v3->unregisterExternalPagesCallbacks(ns_id);
         break;
     }