--- a/dbms/src/Storages/Page/V3/PageDirectory.cpp
+++ b/dbms/src/Storages/Page/V3/PageDirectory.cpp
@@ -1235,6 +1235,7 @@ void PageDirectory::gcApply(PageEntriesEdit && migrated_edit, const WriteLimiter
     LOG_FMT_INFO(log, "GC apply done. [edit size={}]", migrated_edit.size());
 }
 
+
 std::set<PageId> PageDirectory::getAliveExternalIds(NamespaceId ns_id) const
 {
     std::set<PageId> valid_external_ids;
@@ -1248,6 +1249,34 @@ std::set<PageId> PageDirectory::getAliveExternalIds(NamespaceId ns_id) const
             {
                 if (holder->high == ns_id)
                     valid_external_ids.emplace(holder->low);
+                ++iter;
+            }
+        }
+    }
+    return valid_external_ids;
+}
+
+std::map<NamespaceId, std::set<PageId>> PageDirectory::getAliveExternalIds() const
+{
+    std::map<NamespaceId, std::set<PageId>> valid_external_ids;
+    {
+        std::lock_guard guard(external_ids_mutex);
+        for (auto iter = external_ids.begin(); iter != external_ids.end(); /*empty*/)
+        {
+            if (auto holder = iter->lock(); holder == nullptr)
+                iter = external_ids.erase(iter);
+            else
+            {
+                auto ns_id = holder->high;
+                if (auto iter = valid_external_ids.find(ns_id); iter != valid_external_ids.end())
+                {
+                    iter->second.emplace(holder->low);
+                }
+                else
+                {
+                    // create a new set for this ns_id
+                    valid_external_ids.emplace(ns_id, std::set<PageId>{holder->low});
+                }
                 ++iter;
             }
         }