--- a/dbms/src/Storages/Page/V3/PageDirectory.cpp
+++ b/dbms/src/Storages/Page/V3/PageDirectory.cpp
@@ -1170,8 +1170,7 @@ void PageDirectory::apply(PageEntriesEdit && edit, const WriteLimiterPtr & write
                 {
                     // put the new created holder into `external_ids`
                     *holder = r.page_id;
-                    std::lock_guard guard(external_ids_mutex);
-                    external_ids.emplace_back(std::weak_ptr<PageIdV3Internal>(holder));
+                    external_ids_by_ns.addExternalId(holder);
                 }
                 break;
             }
@@ -1235,38 +1234,9 @@ void PageDirectory::gcApply(PageEntriesEdit && migrated_edit, const WriteLimiter
     LOG_FMT_INFO(log, "GC apply done. [edit size={}]", migrated_edit.size());
 }
 
-std::map<NamespaceId, std::set<PageId>> PageDirectory::getAliveExternalIds() const
+std::set<PageId> PageDirectory::getAliveExternalIds(NamespaceId ns_id) const
 {
-    std::map<NamespaceId, std::set<PageId>> valid_external_ids;
-    {
-        std::lock_guard guard(external_ids_mutex);
-        for (auto iter = external_ids.begin(); iter != external_ids.end(); /*empty*/)
-        {
-            if (auto holder = iter->lock(); holder == nullptr)
-            {
-                // the external id is removed from `PageDirectory`,
-                // cleanup the invalid weak_ptr
-                iter = external_ids.erase(iter);
-                continue;
-            }
-            else
-            {
-                const auto ns_id = holder->high;
-                if (auto iter = valid_external_ids.find(ns_id); iter != valid_external_ids.end())
-                {
-                    // insert the valid external id for this ns_id
-                    iter->second.emplace(holder->low);
-                }
-                else
-                {
-                    // create a new set for this ns_id
-                    valid_external_ids.emplace(ns_id, std::set<PageId>{holder->low});
-                }
-                ++iter;
-            }
-        }
-    }
-    return valid_external_ids;
+    return external_ids_by_ns.getAliveIds(ns_id);
 }
 
 std::pair<std::map<BlobFileId, PageIdAndVersionedEntries>, PageSize>