--- a/dbms/src/Storages/Page/V3/PageDirectory.h
+++ b/dbms/src/Storages/Page/V3/PageDirectory.h
@@ -32,6 +32,7 @@
 #include <mutex>
 #include <shared_mutex>
 #include <unordered_map>
+
 #include "Storages/Page/PageDefines.h"
 
 namespace CurrentMetrics