--- a/dbms/src/Storages/Page/V3/PageDirectory/ExternalIdsByNamespace.h
+++ b/dbms/src/Storages/Page/V3/PageDirectory/ExternalIdsByNamespace.h
@@ -1,3 +1,17 @@
+// Copyright 2022 PingCAP, Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 #pragma once
 
 #include <Common/nocopyable.h>
@@ -6,30 +20,32 @@
 namespace DB::PS::V3
 {
 
-struct ExternalIds
-{
-    std::mutex mu;
-    std::list<std::weak_ptr<PageIdV3Internal>> external_ids;
-
-    ExternalIds() = default;
-    ExternalIds(ExternalIds &&rhs): external_ids(std::move(rhs.external_ids)) {}
-};
-
+// A thread-safe class to manage external ids.
+// Manage all external ids by NamespaceId.
 class ExternalIdsByNamespace
 {
 public:
     ExternalIdsByNamespace() = default;
 
+    // Add a external ids
     void addExternalId(const std::shared_ptr<PageIdV3Internal> & external_id);
+    // non thread-safe version, only for restore
     void addExternalIdUnlock(const std::shared_ptr<PageIdV3Internal> & external_id);
 
+    // Get all alive external ids of given `ns_id`
+    // Will also cleanup the invalid external ids.
     std::set<PageId> getAliveIds(NamespaceId ns_id) const;
 
+    // After table dropped, the `getAliveIds` with specified
+    // `ns_id` will not be cleaned. We need this method to
+    // cleanup all external id ptrs.
+    void unregisterNamespace(NamespaceId ns_id);
 
     DISALLOW_COPY_AND_MOVE(ExternalIdsByNamespace);
 
 private:
     mutable std::mutex mu;
+    using ExternalIds = std::list<std::weak_ptr<PageIdV3Internal>>;
     using StableMap = std::map<NamespaceId, ExternalIds>;
     mutable StableMap ids_by_ns;
 };