--- a/dbms/src/Storages/Page/V3/PageDirectoryFactory.cpp
+++ b/dbms/src/Storages/Page/V3/PageDirectoryFactory.cpp
@@ -149,7 +149,7 @@ void PageDirectoryFactory::applyRecord(
             if (holder)
             {
                 *holder = r.page_id;
-                dir->external_ids.emplace_back(std::weak_ptr<PageIdV3Internal>(holder));
+                dir->external_ids_by_ns.addExternalIdUnlock(holder);
             }
             break;
         }
@@ -162,7 +162,7 @@ void PageDirectoryFactory::applyRecord(
             if (holder)
             {
                 *holder = r.page_id;
-                dir->external_ids.emplace_back(std::weak_ptr<PageIdV3Internal>(holder));
+                dir->external_ids_by_ns.addExternalId(holder);
             }
             break;
         }