### What problem does this PR solve?

Issue Number: ref https://github.com/pingcap/tiflash/issues/5697

Problem Summary:
The GC total time in logging does not include the time costs in `clean_external_page`.

### What is changed and how it works?

* Add a struct `GCTimeStatistics` to handle logging
* Add the time cost in `cleanExternalPage` into GC total time
* Add `ExternalIdsByNamespace` to
  * Manage external ids by namespace
  * Shorten the GC time of cleaning external pages from 2500+ms to 420ms with about 5000 empty tables

```
# before
[2022/08/31 13:06:27.762 +08:00] [DEBUG] [PageStorageImpl.cpp:330] ["PageStorage:__global__.data GC finished without full gc. [total time=2514ms] [dump snapshots=0ms] [gc in mem entries=1ms] [blobstore remove entries=0ms] [blobstore get status=0ms] [get gc entries=0ms] [blobstore full gc=0ms] [gc apply=0ms] [external callbacks=4933] [external gc=2512ms] [get alive=2098.75ms] [scanner=408.65ms] [remover=4.90ms]"] [thread_id=163]

# after
[2022/08/31 13:10:22.715 +08:00] [DEBUG] [PageStorageImpl.cpp:330] ["PageStorage:__global__.data GC finished without full gc. [total time=427ms] [dump snapshots=0ms] [gc in mem entries=1ms] [blobstore remove entries=0ms] [blobstore get status=0ms] [get gc entries=0ms] [blobstore full gc=0ms] [gc apply=0ms] [external callbacks=4933] [external gc=425ms] [get alive=14.70ms] [scanner=407.08ms] [remover=3.83ms]"] [thread_id=136]
```

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```